api/requirements.txt
fastapi[standard]==0.116.1
kafka-python==2.2.15
uvicorn[standard]==0.35.0
websocket-client==1.8.0
python-dotenv==1.0.0
pyyaml==6.0.1
